Updates to v1.9.1.2

+++ src/Logger/ListTable.php
@@ -118,7 +118,7 @@ class ListTable extends WP_List_Table {
 			[
 				'total_items' => $total_items,
 				'per_page'    => $per_page,
-				'total_pages' => ceil( $total_items / $per_page ),
+				'total_pages' => (int) ceil( $total_items / $per_page ),
 			]
 		);
 	}
